pyproject.toml
[tool.poetry]
name = "bmg"
version = "0.1.0"
description = "BMG Technical Test"
authors = ["raedi <raedinurdiansyah@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
Flask = "^2.0.1"
SQLAlchemy = "^1.4.23"
marshmallow = "^3.13.0"
Flask-SQLAlchemy = "^2.5.1"
Flask-Migrate = "^3.1.0"
flask-marshmallow = "^0.14.0"
PyJWT = "^2.1.0"
python-dotenv = "^0.19.0"
marshmallow-sqlalchemy = "^0.26.1"
alembic = "^1.7.3"
psycopg2-binary = "^2.9.1"
Werkzeug = "^2.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
